bugfix: VariableChange for oops compatibility
[oops PR#1098](https://github.com/JCSDA-internal/oops/pull/1098)
introduced a requirement for a default variable change class.  The
implementation here is a stripped down version of the opsinputs class of
the same name.

Introducing this class depends upon the ability to assign one state to
another. As a temporary fix the VariableChange class does nothing and
the state `=` operator merely copies across the validity time of a state.