bugfix: fix state operator= to copy state
It seems that a variableChange happens regardless of whether or not any
variables need changing. In this process the state is set by assignment
and the associated data ought to be copied across. This probably ought
to happen via a shallow copy but a deep copy is OK for now.