tests: move config to dedicated file and revert
Revert the change made to ostia_seaice_obs.yaml. I did this because
this file is the only example of orcamodelHofX.x at the moment.

I moved the test data to a new yaml file. We cannot add a test for this
until atlas is updated to the latest version in jopa-bundle.