Fix the geovals vector ordering for profiles (#13)
The "results" vector that contains the GeoVaLs data output by the interpolator was defined in the wrong order. This was only an issue if was more than one level in the data, as there is for profile data.

The ordering should be by variable, z-level, and then horizontal location, with the location varying the fastest.